fix: update dtype in batch processing script to use Float16 for Ampere GPUs

+++ scripts/batch_process_kaggle.py
@@ -191,8 +191,9 @@ def main():
     )
 
     # Enable optimizations for Ampere GPUs (RTX 3000+, A100, etc.)
+    # Note: Grounding DINO's CUDA extension doesn't support BFloat16, so we use Float16 instead
     if torch.cuda.is_available():
-        torch.autocast(device_type=device, dtype=torch.bfloat16).__enter__()
+        torch.autocast(device_type=device, dtype=torch.float16).__enter__()
         if torch.cuda.get_device_properties(0).major >= 8:
             torch.backends.cuda.matmul.allow_tf32 = True
             torch.backends.cudnn.allow_tf32 = True
